workers/cls-annotation/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.24.0
transformers>=4.40.0
Pillow>=10.0.0
requests>=2.31.0
sentencepiece>=0.2.0
protobuf>=4.25.0
runpod>=1.6.0
